Added handler boolean + other things
Added a Handler Boolean to execute tasks once, added a second logcat value to check for, for when the default assistant is not google. 

`service.performGlobalAction(AccessibilityService.GLOBAL_ACTION_BACK) is some times not triggered so there is a if/else added to dispatch the back button (not not the custom action) in case the back button does not dispatch properly first time. Leaving the google assistant on screen (this only effects MediaKeyActions) hence the `customAction is MediaKeyAction` condition.