FlipToShush: Overhaul the inner workings
Now proximity will be the first check that needs to
be passed, followed by checks for device position with
accelerometer.

With this the requirements of wakeup accelerometer
sensor is removed. Although battery consumption could
be higher, because partial wakelock will be acquired
when accelerometer data is needed. Devices with wakeup
accelerometer sensor will not use any wakelock.

This commit also tightens the trigger point of DND
switching to avoid accidental activation on pocket.